revert(llm-proxy): restore the LLM Proxies list page and navigation
Keeps the Connect-page changes (Model Router tab, terminal-card endpoint
and MCP sections); reverts the singular workspace at /llm/proxies, the
manage subroute, nav rename, virtual-keys create deep link, and the
workspace e2e coverage back to the main baseline.
