--- platform/frontend/src/app/llm/proxies/manage/page.client.tsx
@@ -1,561 +1,0 @@
-"use client";
-
-import { type archestraApiTypes, E2eTestId } from "@archestra/shared";
-import type { ColumnDef, SortingState } from "@tanstack/react-table";
-import { ChevronDown, ChevronUp, Plus } from "lucide-react";
-import Link from "next/link";
-import { useRouter } from "next/navigation";
-import { useCallback, useEffect, useState } from "react";
-import { toast } from "sonner";
-import { ErrorBoundary } from "@/app/_parts/error-boundary";
-import { AgentDialog } from "@/components/agent-dialog";
-import { AgentIcon } from "@/components/agent-icon";
-import { AgentNameCell } from "@/components/agent-name-cell";
-import {
-  ActiveFilterBadges,
-  AgentDeletedStatusFilter,
-  AgentScopeFilter,
-} from "@/components/agent-scope-filter";
-import { CloneAgentDialog } from "@/components/clone-agent-dialog";
-import { DeleteConfirmDialog } from "@/components/delete-confirm-dialog";
-import { ExternalDocsLink } from "@/components/external-docs-link";
-import { LoadingSpinner, LoadingWrapper } from "@/components/loading";
-import { PageLayout } from "@/components/page-layout";
-import { PermissionRequirementHint } from "@/components/permission-requirement-hint";
-import { PostCreateConnectDialog } from "@/components/post-create-connect-dialog";
-import { ResourceVisibilityBadge } from "@/components/resource-visibility-badge";
-import { SearchInput } from "@/components/search-input";
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import { DataTable } from "@/components/ui/data-table";
-import { PermissionButton } from "@/components/ui/permission-button";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
-import { DEFAULT_SORT_BY, DEFAULT_SORT_DIRECTION } from "@/consts";
-import {
-  useDeleteProfile,
-  useProfile,
-  useProfilesPaginated,
-  useRestoreProfile,
-} from "@/lib/agent.query";
-import { useHasPermissions, useSession } from "@/lib/auth/auth.query";
-import { getFrontendDocsUrl } from "@/lib/docs/docs";
-import { useDataTableQueryParams } from "@/lib/hooks/use-data-table-query-params";
-import { useDialogUrlParam } from "@/lib/hooks/use-dialog-url-param";
-import { useMyTeams } from "@/lib/teams/team.query";
-import { LlmProxyActions } from "./llm-proxy-actions";
-
-type LlmProxiesInitialData = {
-  agents: archestraApiTypes.GetAgentsResponses["200"] | null;
-  teams: archestraApiTypes.GetTeamsResponses["200"]["data"];
-};
-
-export default function LlmProxiesPage({
-  initialData,
-}: {
-  initialData?: LlmProxiesInitialData;
-}) {
-  return (
-    <div className="w-full h-full">
-      <ErrorBoundary>
-        <LlmProxies initialData={initialData} />
-      </ErrorBoundary>
-    </div>
-  );
-}
-
-function SortIcon({
-  isSorted,
-}: {
-  isSorted:
-    | NonNullable<archestraApiTypes.GetAgentsData["query"]>["sortDirection"]
-    | false;
-}) {
-  const upArrow = <ChevronUp className="h-3 w-3" />;
-  const downArrow = <ChevronDown className="h-3 w-3" />;
-  if (isSorted === "asc") {
-    return upArrow;
-  }
-  if (isSorted === "desc") {
-    return downArrow;
-  }
-  return (
-    <div className="text-muted-foreground flex flex-col items-center">
-      {upArrow}
-      <span className="mt-[-4px]">{downArrow}</span>
-    </div>
-  );
-}
-
-function LlmProxies({ initialData }: { initialData?: LlmProxiesInitialData }) {
-  const docsUrl = getFrontendDocsUrl("platform-llm-proxy");
-  const {
-    searchParams,
-    pageIndex,
-    pageSize,
-    offset,
-    updateQueryParams,
-    setPagination,
-  } = useDataTableQueryParams();
-
-  const nameFilter = searchParams.get("name") || "";
-  const sortByFromUrl = searchParams.get("sortBy") as
-    | "name"
-    | "createdAt"
-    | "toolsCount"
-    | "team"
-    | null;
-  const sortDirectionFromUrl = searchParams.get("sortDirection") as
-    | "asc"
-    | "desc"
-    | null;
-  const scopeFromUrl = searchParams.get("scope") as
-    | "personal"
-    | "team"
-    | "org"
-    | "built_in"
-    | null;
-  const teamIdsFromUrl = searchParams.get("teamIds");
-  const authorIdsFromUrl = searchParams.get("authorIds");
-  const excludeAuthorIdsFromUrl = searchParams.get("excludeAuthorIds");
-  const labelsFromUrl = searchParams.get("labels");
-  const statusFromUrl = searchParams.get("status") as
-    | "active"
-    | "deleted"
-    | null;
-  const isDeletedView = statusFromUrl === "deleted";
-
-  const sortBy = sortByFromUrl || DEFAULT_SORT_BY;
-  const sortDirection = sortDirectionFromUrl || DEFAULT_SORT_DIRECTION;
-  const { data: canDeleteAgents } = useHasPermissions({ agent: ["delete"] });
-  const proxyAgentTypes: Array<"llm_proxy" | "profile"> =
-    isDeletedView && !canDeleteAgents
-      ? ["llm_proxy"]
-      : ["llm_proxy", "profile"];
-
-  const { data: agentsResponse, isPending } = useProfilesPaginated({
-    initialData: initialData?.agents ?? undefined,
-    limit: pageSize,
-    offset,
-    sortBy,
-    sortDirection,
-    name: nameFilter || undefined,
-    agentTypes: proxyAgentTypes,
-    scope: scopeFromUrl || undefined,
-    teamIds: teamIdsFromUrl ? teamIdsFromUrl.split(",") : undefined,
-    authorIds: authorIdsFromUrl ? authorIdsFromUrl.split(",") : undefined,
-    excludeAuthorIds: excludeAuthorIdsFromUrl
-      ? excludeAuthorIdsFromUrl.split(",")
-      : undefined,
-    excludeOtherPersonalAgents:
-      scopeFromUrl !== "personal" &&
-      !authorIdsFromUrl &&
-      !excludeAuthorIdsFromUrl
-        ? true
-        : undefined,
-    labels: labelsFromUrl || undefined,
-    status: statusFromUrl || undefined,
-  });
-  const { data: canReadTeams } = useHasPermissions({ team: ["read"] });
-
-  const { data: userTeams } = useMyTeams({
-    enabled: !!canReadTeams,
-  });
-
-  const { data: isAdmin } = useHasPermissions({ llmProxy: ["admin"] });
-  const { data: isTeamAdmin } = useHasPermissions({
-    llmProxy: ["team-admin"],
-  });
-  const { data: session } = useSession();
-  const currentUserId = session?.user?.id;
-  const userTeamIdSet = new Set((userTeams ?? []).map((t) => t.id));
-
-  const [sorting, setSorting] = useState<SortingState>([
-    { id: sortBy, desc: sortDirection === "desc" },
-  ]);
-
-  useEffect(() => {
-    setSorting([{ id: sortBy, desc: sortDirection === "desc" }]);
-  }, [sortBy, sortDirection]);
-
-  type ProxyData = archestraApiTypes.GetAgentsResponses["200"]["data"][number];
-
-  const router = useRouter();
-
-  const [isCreateDialogOpen, setIsCreateDialogOpen] = useState(false);
-  const [postCreateProxy, setPostCreateProxy] = useState<{
-    id: string;
-    name: string;
-  } | null>(null);
-  const navigateToConnection = useCallback(
-    (agentId: string) => {
-      router.push(
-        `/connection?proxyId=${encodeURIComponent(agentId)}&from=table`,
-      );
-    },
-    [router],
-  );
-  const editId = searchParams.get("edit");
-  const { data: editFromUrl } = useProfile(editId ?? undefined);
-  const editDialog = useDialogUrlParam<ProxyData>({
-    paramName: "edit",
-    entityFromUrl: editFromUrl ?? null,
-  });
-  const [deletingProxyId, setDeletingProxyId] = useState<string | null>(null);
-  const [cloningProxy, setCloningProxy] = useState<ProxyData | null>(null);
-  const restoreProxy = useRestoreProfile();
-
-  const handleSortingChange = useCallback(
-    (updater: SortingState | ((old: SortingState) => SortingState)) => {
-      const newSorting =
-        typeof updater === "function" ? updater(sorting) : updater;
-      setSorting(newSorting);
-
-      if (newSorting.length > 0) {
-        updateQueryParams({
-          page: "1",
-          sortBy: newSorting[0].id,
-          sortDirection: newSorting[0].desc ? "desc" : "asc",
-        });
-      } else {
-        updateQueryParams({
-          page: "1",
-          sortBy: null,
-          sortDirection: null,
-        });
-      }
-    },
-    [sorting, updateQueryParams],
-  );
-
-  const handlePaginationChange = useCallback(
-    (newPagination: { pageIndex: number; pageSize: number }) => {
-      setPagination(newPagination);
-    },
-    [setPagination],
-  );
-
-  const agents = agentsResponse?.data || [];
-  const pagination = agentsResponse?.pagination;
-  const showLoading = isPending && !initialData?.agents;
-
-  const columns: ColumnDef<ProxyData>[] = [
-    {
-      id: "icon",
-      size: 40,
-      enableSorting: false,
-      header: "",
-      cell: ({ row }) => (
-        <div className="flex items-center justify-center">
-          <AgentIcon
-            icon={row.original.icon}
-            size={20}
-            fallbackType="llm_proxy"
-          />
-        </div>
-      ),
-    },
-    {
-      id: "name",
-      accessorKey: "name",
-      size: 240,
-      header: ({ column }) => (
-        <Button
-          variant="ghost"
-          className="h-auto !p-0 font-medium hover:bg-transparent"
-          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-        >
-          Name
-          <SortIcon isSorted={column.getIsSorted()} />
-        </Button>
-      ),
-      cell: ({ row }) => {
-        const agent = row.original;
-        return (
-          <AgentNameCell
-            name={agent.name}
-            description={agent.description}
-            extraBadges={
-              agent.agentType === "profile" ? (
-                <TooltipProvider>
-                  <Tooltip>
-                    <TooltipTrigger asChild>
-                      <Badge
-                        variant="outline"
-                        className="bg-orange-500/10 text-orange-600 border-orange-500/30 text-xs cursor-help"
-                      >
-                        Profile
-                      </Badge>
-                    </TooltipTrigger>
-                    <TooltipContent>
-                      This is a legacy entity that works both as MCP Gateway and
-                      LLM Proxy
-                    </TooltipContent>
-                  </Tooltip>
-                </TooltipProvider>
-              ) : null
-            }
-            labels={agent.labels}
-          />
-        );
-      },
-    },
-    {
-      id: "team",
-      header: "Accessible to",
-      enableSorting: false,
-      cell: ({ row }) => (
-        <ResourceVisibilityBadge
-          scope={row.original.scope}
-          teams={row.original.teams}
-          authorId={row.original.authorId}
-          authorName={row.original.authorName}
-          currentUserId={currentUserId}
-          showSelfAsMe
-        />
-      ),
-    },
-    {
-      id: "actions",
-      header: "Actions",
-      enableHiding: false,
-      cell: ({ row }) => {
-        const agent = row.original;
-        const scope = agent.scope;
-        const authorId = agent.authorId;
-        const agentTeams = agent.teams;
-        const isPersonal = scope === "personal";
-        const isTeamScoped = scope === "team";
-        const isOwner = !!currentUserId && authorId === currentUserId;
-        const isMemberOfAgentTeam = agentTeams?.some((t) =>
-          userTeamIdSet.has(t.id),
-        );
-        const canModify =
-          !!isAdmin ||
-          (isTeamScoped && !!isTeamAdmin && !!isMemberOfAgentTeam) ||
-          (isPersonal && isOwner);
-        return (
-          <LlmProxyActions
-            agent={agent}
-            canModify={canModify}
-            onConnect={(a) => navigateToConnection(a.id)}
-            onEdit={editDialog.open}
-            onDelete={setDeletingProxyId}
-            onRestore={(agentId) => {
-              restoreProxy.mutate(agentId, {
-                onSuccess: (data) => {
-                  if (!data) return;
-                  toast.success("LLM Proxy restored successfully");
-                },
-              });
-            }}
-            onClone={setCloningProxy}
-          />
-        );
-      },
-    },
-  ];
-
-  return (
-    <LoadingWrapper
-      isPending={showLoading}
-      loadingFallback={<LoadingSpinner />}
-    >
-      <PageLayout
-        title="All LLM Proxies"
-        description={
-          <p className="text-sm text-muted-foreground">
-            Create, clone, archive, and restore proxies.{" "}
-            <Link
-              href="/llm/proxies"
-              className="underline hover:text-foreground"
-            >
-              Back to the LLM Proxy workspace
-            </Link>
-            {docsUrl && (
-              <>
-                {" · "}
-                <ExternalDocsLink
-                  href={docsUrl}
-                  className="underline hover:text-foreground"
-                  showIcon={false}
-                >
-                  Docs
-                </ExternalDocsLink>
-              </>
-            )}
-          </p>
-        }
-        actionButton={
-          <PermissionButton
-            permissions={{ llmProxy: ["create"] }}
-            onClick={() => setIsCreateDialogOpen(true)}
-            data-testid={E2eTestId.CreateAgentButton}
-          >
-            <Plus className="h-4 w-4" />
-            Create LLM Proxy
-          </PermissionButton>
-        }
-      >
-        <div>
-          <div>
-            <div className="mb-6 flex flex-col gap-2">
-              <div className="flex items-center gap-4">
-                <SearchInput
-                  objectNamePlural="proxies"
-                  searchFields={["name"]}
-                  paramName="name"
-                />
-                <AgentScopeFilter
-                  ownerLabelPlural="LLM proxies"
-                  adminPermission={{ llmProxy: ["admin"] }}
-                />
-                <AgentDeletedStatusFilter
-                  deletePermission={{ llmProxy: ["delete"] }}
-                />
-              </div>
-              {!canReadTeams && (
-                <PermissionRequirementHint
-                  message="Team-based filters and sharing details are unavailable without"
-                  permissions={[{ resource: "team", action: "read" }]}
-                />
-              )}
-              <ActiveFilterBadges adminPermission={{ llmProxy: ["admin"] }} />
-            </div>
-
-            <div data-testid={E2eTestId.AgentsTable}>
-              <DataTable
-                columns={columns}
-                data={agents}
-                sorting={sorting}
-                onSortingChange={handleSortingChange}
-                manualSorting={true}
-                manualPagination={true}
-                pagination={{
-                  pageIndex,
-                  pageSize,
-                  total: pagination?.total ?? 0,
-                }}
-                onPaginationChange={handlePaginationChange}
-                hasActiveFilters={Boolean(
-                  nameFilter ||
-                    scopeFromUrl ||
-                    teamIdsFromUrl ||
-                    authorIdsFromUrl ||
-                    excludeAuthorIdsFromUrl ||
-                    labelsFromUrl ||
-                    isDeletedView,
-                )}
-                onClearFilters={() =>
-                  updateQueryParams({
-                    name: null,
-                    scope: null,
-                    teamIds: null,
-                    authorIds: null,
-                    excludeAuthorIds: null,
-                    labels: null,
-                    status: null,
-                    page: "1",
-                  })
-                }
-                emptyMessage={
-                  isDeletedView
-                    ? "No deleted LLM proxies found"
-                    : "No LLM proxies found"
-                }
-                filteredEmptyMessage={
-                  isDeletedView
-                    ? "No deleted LLM proxies found."
-                    : "No LLM proxies match your filters. Try adjusting your search."
-                }
-              />
-            </div>
-
-            <AgentDialog
-              open={isCreateDialogOpen}
-              onOpenChange={setIsCreateDialogOpen}
-              agentType="llm_proxy"
-              defaultIconType="llm_proxy"
-              onCreated={(created) => {
-                setIsCreateDialogOpen(false);
-                setPostCreateProxy(created);
-              }}
-            />
-
-            <PostCreateConnectDialog
-              created={postCreateProxy}
-              agentType="llm_proxy"
-              onOpenChange={(open) => {
-                if (!open) setPostCreateProxy(null);
-              }}
-            />
-
-            <AgentDialog
-              open={!!editDialog.entity}
-              onOpenChange={(open) => !open && editDialog.close()}
-              agent={editDialog.entity}
-              agentType={editDialog.entity?.agentType || "llm_proxy"}
-              defaultIconType="llm_proxy"
-            />
-
-            {deletingProxyId && (
-              <DeleteProxyDialog
-                agentId={deletingProxyId}
-                open={!!deletingProxyId}
-                onOpenChange={(open) => !open && setDeletingProxyId(null)}
-              />
-            )}
-
-            <CloneAgentDialog
-              agent={cloningProxy}
-              onOpenChange={(open) => {
-                if (!open) setCloningProxy(null);
-              }}
-              onCloned={(cloned) => {
-                // Open edit dialog for the clone so user can rename immediately
-                editDialog.open(cloned as ProxyData);
-              }}
-            />
-          </div>
-        </div>
-      </PageLayout>
-    </LoadingWrapper>
-  );
-}
-
-function DeleteProxyDialog({
-  agentId,
-  open,
-  onOpenChange,
-}: {
-  agentId: string;
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-}) {
-  const deleteProxy = useDeleteProfile();
-
-  const handleDelete = useCallback(async () => {
-    const result = await deleteProxy.mutateAsync(agentId);
-    if (result) {
-      toast.success("LLM Proxy deleted successfully");
-      onOpenChange(false);
-    }
-  }, [agentId, deleteProxy, onOpenChange]);
-
-  return (
-    <DeleteConfirmDialog
-      open={open}
-      onOpenChange={onOpenChange}
-      title="Delete LLM Proxy"
-      description="Are you sure you want to delete this LLM Proxy? This action cannot be undone."
-      isPending={deleteProxy.isPending}
-      onConfirm={handleDelete}
-      confirmLabel="Delete LLM Proxy"
-      pendingLabel="Deleting..."
-    />
-  );
-}
